requirements.txt
pillow
opencv-python
numpy
matplotlib
tensorflow
torch
torchvision
torchmetrics
kaggle
pandas
jupyter
ipykernel